requirements.txt
appdirs==1.4.3
gunicorn==19.7.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
Django==1.10
django-storages==1.5.2
olefile==0.44
Pillow==4.0.0
psycopg2==2.7
python-dateutil==2.6.0
pytz==2016.10
Unidecode==0.4.20
virtualenv==15.1.0
django-crispy-forms
django-smartfields
python3-memcached==1.51
boto==2.46.1
boto3==1.4.4
django-import-export==0.5.1